+++ R/info.R
@@ -1,7 +1,0 @@
-
-# Inform users about the breaking changes of v0.5 compared to previous versions
-.onAttach <- function(libname, pkgname) {
- packageStartupMessage(paste("Welcome to tidystats 0.5.\nPlease note that this",
-   "update (again) breaks much of the code from previous versions; so please",
-   "check your old code carefully."))
-}